% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ipl.R
\docType{package}
\name{Ipl}
\alias{Ipl-package}
\alias{Ipl}
\title{Ipl: A Comprehensive IPL Data Analysis Package}
\description{
The \code{Ipl} package provides tools for analyzing key performance metrics of IPL players,
focusing on metrics relevant to T20 cricket strategies. It includes a Shiny app for
visualizing and filtering players based on batting and bowling performance data.
The package also provides a dataset, \code{ipl_high_performers}, with data on
top IPL players, which users can explore to make data-driven decisions for team building.
}
\section{Key Features}{

\itemize{
\item \strong{Shiny App}: Interactive app to explore IPL player data with filters for team and player type.
\item \strong{Dataset}: \code{ipl_high_performers} includes batting and bowling statistics for high-performing players.
\item \strong{Batting Analysis}: Scatter plot visualization for \code{Batting Average} vs. \code{Strike Rate}.
\item \strong{Bowling Analysis}: Scatter plot visualization for \code{Bowling Average} vs. \code{Economy Rate}.
}
}

\section{Installation}{

The package can be installed from GitHub using \code{remotes::install_github("yourusername/Ipl")}.
}

\section{Example}{

After loading the package, launch the Shiny app to explore the player performance data:
\code{launch_app()}
}

\section{Datasets}{

\itemize{
\item \code{ipl_high_performers}: IPL players' performance metrics for batting and bowling.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://etc5523-2024.github.io/assignment-4-packages-and-shiny-apps-shiveshkrishna/}
}

}
\author{
\strong{Maintainer}: Shivesh Krishna Palanisamy \email{spal0044@student.monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
